fix: moved tenant configs to namespace ionos\essentials\tenants\config
